fix: fix some url bugs
